pyproject.toml
[project]
name = "rag-pdf-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flagembedding>=1.3.5",
    "huggingface-hub>=1.4.1",
    "langchain>=1.2.10",
    "numpy>=2.4.2",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
